+++ source/digits_hits/src/GateToBinary.cc
@@ -184,12 +184,11 @@ void GateToBinary::RecordEndOfEvent( G4Event const* event )
 
   if( m_outFileHitsFlag )
     {
-      //GateHitsCollection* CHC = GetOutputMgr()->
-       // GetHitCollection();
+
 	  //OK GND 2022
 	  std::vector<GateHitsCollection*> CHC_vector = GetOutputMgr()->GetHitCollections();
 
-	 for (long unsigned int i=0; i<CHC_vector.size();i++ )//HC_vector.size()
+	 for (size_t i=0; i<CHC_vector.size();i++ )//HC_vector.size()
 		{
 		 GateHitsCollection* CHC = CHC_vector[i];
